Safari-611.1.10/LayoutTests/fast/repaint/overflow-hidden-with-self-painting-child-layer-expected.txt

PASS window.internals.repaintRectsAsText().indexOf('200 200') is not -1
PASS successfullyParsed is true

TEST COMPLETE
